# docs/openapi/gsf-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GSF API",
    "version": "0.1.1",
    "description": "Public HTTP surface of GSF, served by the Next.js route handlers. Generated by `pnpm openapi` — do not edit by hand."
  },
  "paths": {
...
    "/api/exploration/edges": {
      "get": {
        "description": "explorationApi.getEdges — table connections backed by a shared SQL query or a foreign key.",
        "operationId": "get_api_exploration_edges",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_ExplorationEdge_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Data Exploration Edges",
        "tags": ["exploration"],
        "x-gsf-source": "frontend/app/api/exploration/edges/route.ts",
        "x-gsf-permissions": {
          "catalog": ["read"]
        },
        "x-gsf-proxied": true
      }
    },
    "/api/exploration/graph": {
      "get": {
        "description": "explorationApi.getDataExplorationGraph — full data-layer graph ({nodes, links}), zone-scoped for viewers.",
        "operationId": "get_api_exploration_graph",
        "parameters": [
          {
            "description": "Max nodes returned; always capped at 500 server-side.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "description": "Max nodes returned; always capped at 500 server-side.",
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_DataExplorationGraph_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Data Exploration Graph",
        "tags": ["exploration"],
        "x-gsf-source": "frontend/app/api/exploration/graph/route.ts",
        "x-gsf-permissions": {
          "catalog": ["read"]
        },
        "x-gsf-proxied": true
      }
    },
    "/api/exploration/nodes/{node_id}/relationships": {
      "get": {
        "description": "explorationApi.getRelatedNodes — a zone-scoped page of related Exploration nodes.",
        "operationId": "get_api_exploration_nodes_node_id_relationships",
        "parameters": [
          {
            "description": "Table id when *layer* is ``data``, Term id when it is ``semantic``.",
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "description": "Table id when *layer* is ``data``, Term id when it is ``semantic``.",
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "description": "Which graph to walk: ``data`` relates tables sharing a SQL query or a foreign key, ``semantic`` relates terms sharing a table.",
            "in": "query",
            "name": "layer",
            "required": true,
            "schema": {
              "description": "Which graph to walk: ``data`` relates tables sharing a SQL query or a foreign key, ``semantic`` relates terms sharing a table.",
              "enum": ["data", "semantic"],
              "title": "Layer",
              "type": "string"
            }
          },
          {
            "description": "Rows to skip before the page starts.",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip before the page starts.",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "Page size, at most 100. Omit to return every matching row in one response.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page size, at most 100. Omit to return every matching row in one response.",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_ExplorationRelatedNodes_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Exploration Node Relationships",
        "tags": ["exploration"],
        "x-gsf-source": "frontend/app/api/exploration/nodes/[node_id]/relationships/route.ts",
        "x-gsf-permissions": {
          "catalog": ["read"]
        },
        "x-gsf-proxied": true
      }
    },
    "/api/exploration/semantic-graph": {
      "get": {
        "description": "explorationApi.getSemanticExplorationGraph — full semantic-layer graph ({nodes, links}), zone-scoped for viewers.",
        "operationId": "get_api_exploration_semantic_graph",
        "parameters": [
          {
            "description": "Max nodes returned; always capped at 500 server-side.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "description": "Max nodes returned; always capped at 500 server-side.",
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_SemanticExplorationGraph_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Semantic Exploration Graph",
        "tags": ["exploration"],
        "x-gsf-source": "frontend/app/api/exploration/semantic-graph/route.ts",
        "x-gsf-permissions": {
          "catalog": ["read"]
        },
        "x-gsf-proxied": true
      }
    },
    "/api/exploration/tables/{table_id}/details": {
      "get": {
        "description": "explorationApi.getTableExplorationDetails — columns-adjacent SQL and Term details for one table.",
        "operationId": "get_api_exploration_tables_table_id_details",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          },
          {
            "description": "Rows to skip before the page starts.",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip before the page starts.",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "Page size, at most 100. Omit to return every matching row in one response.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page size, at most 100. Omit to return every matching row in one response.",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TableExplorationDetails_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Table Exploration Details",
        "tags": ["exploration"],
        "x-gsf-source": "frontend/app/api/exploration/tables/[table_id]/details/route.ts",
        "x-gsf-permissions": {
          "catalog": ["read"]
        },
        "x-gsf-proxied": true
      }
    },
    "/api/exploration/tables/zones": {
      "get": {
        "description": "explorationApi.getTableZonesMap — {table_id: [zone, ...]} map, zone-scoped for viewers.",
        "operationId": "get_api_exploration_tables_zones",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_str__list_ZoneChip___"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Table Zones",
        "tags": ["exploration"],
        "x-gsf-source": "frontend/app/api/exploration/tables/zones/route.ts",
        "x-gsf-permissions": {
          "catalog": ["read"]
        },
        "x-gsf-proxied": true
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataExplorationGraph": {
        "additionalProperties": true,
        "description": "``{nodes, links}`` for the data layer, in one response.",
        "properties": {
          "links": {
            "items": {
              "$ref": "#/components/schemas/ExplorationEdge"
            },
            "title": "Links",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/DataGraphNode"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "title": "DataExplorationGraph",
        "type": "object"
      },
      "DataGraphNode": {
        "additionalProperties": true,
        "description": "A table on the data-layer exploration graph.\n\n``relationship_count`` is the table's degree over the whole accessible\nedge set, so it does not shrink when ``limit`` truncates the payload.",
        "properties": {
          "columns_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns Count"
          },
          "database_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Id"
          },
          "database_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "relationship_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship Count"
          },
          "schema_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Id"
          },
          "schema_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Name"
          },
          "sql_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sql Count"
          },
          "table_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Type"
          },
          "terms_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Count"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/ZoneChip"
            },
            "title": "Zones",
            "type": "array"
          }
        },
        "required": ["id"],
        "title": "DataGraphNode",
        "type": "object"
      },
...
      "DataResponse_DataExplorationGraph_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DataExplorationGraph"
          }
        },
        "required": ["data"],
        "title": "DataResponse[DataExplorationGraph]",
        "type": "object"
      },
      "DataResponse_dict_str__list_ZoneChip___": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ZoneChip"
              },
              "type": "array"
            },
            "title": "Data",
            "type": "object"
          }
        },
        "required": ["data"],
        "title": "DataResponse[dict[str, list[ZoneChip]]]",
        "type": "object"
      },
...
      "DataResponse_ExplorationRelatedNodes_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExplorationRelatedNodes"
          }
        },
        "required": ["data"],
        "title": "DataResponse[ExplorationRelatedNodes]",
        "type": "object"
      },
...
      "DataResponse_SemanticExplorationGraph_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SemanticExplorationGraph"
          }
        },
        "required": ["data"],
        "title": "DataResponse[SemanticExplorationGraph]",
        "type": "object"
      },
...
      "DataResponse_TableExplorationDetails_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TableExplorationDetails"
          }
        },
        "required": ["data"],
        "title": "DataResponse[TableExplorationDetails]",
        "type": "object"
      },
...
      "ExplorationEdge": {
        "additionalProperties": true,
        "description": "Two tables connected by a shared SQL query and/or a foreign key.\n\n``queries`` is empty for an edge that exists only because of a foreign\nkey; ``foreign_keys`` is empty for a SQL-only edge.",
        "properties": {
          "foreign_keys": {
            "items": {
              "$ref": "#/components/schemas/ForeignKeyRef"
            },
            "title": "Foreign Keys",
            "type": "array"
          },
          "queries": {
            "items": {
              "type": "string"
            },
            "title": "Queries",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "via_foreign_key": {
            "default": false,
            "title": "Via Foreign Key",
            "type": "boolean"
          }
        },
        "required": ["source", "target"],
        "title": "ExplorationEdge",
        "type": "object"
      },
      "ExplorationRelatedNode": {
        "additionalProperties": true,
        "description": "One related node. The catalog path fields are data-layer only.",
        "properties": {
          "database_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "relationship_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship Count"
          },
          "schema_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Id"
          },
          "table_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Type"
          }
        },
        "required": ["id"],
        "title": "ExplorationRelatedNode",
        "type": "object"
      },
      "ExplorationRelatedNodes": {
        "additionalProperties": true,
        "description": "One page of related nodes; ``total`` counts every related node.",
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/ExplorationRelatedNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "title": "ExplorationRelatedNodes",
        "type": "object"
      },
...
      "ForeignKeyRef": {
        "additionalProperties": true,
        "description": "One foreign-key column pair behind an exploration edge.\n\n``source_column`` always names a column on the edge's ``source``. The\nsample values come straight off the Column node, unparsed — profiling\nstores them as a JSON string, a catalog PATCH stores a list.",
        "properties": {
          "source_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Column"
          },
          "source_sample_values": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Sample Values"
          },
          "target_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Column"
          },
          "target_sample_values": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Sample Values"
          }
        },
        "title": "ForeignKeyRef",
        "type": "object"
      },
      "GraphLink": {
        "additionalProperties": true,
        "description": "An undirected term↔term link (two terms sharing at least one table).",
        "properties": {
          "source": {
            "title": "Source",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": ["source", "target"],
        "title": "GraphLink",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListResponse_ExplorationEdge_": {
        "additionalProperties": true,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ExplorationEdge"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": ["data", "count"],
        "title": "ListResponse[ExplorationEdge]",
        "type": "object"
      },
...
      "SemanticExplorationGraph": {
        "additionalProperties": true,
        "description": "``{nodes, links}`` for the semantic layer, in one response.",
        "properties": {
          "links": {
            "items": {
              "$ref": "#/components/schemas/GraphLink"
            },
            "title": "Links",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/SemanticGraphNode"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "title": "SemanticExplorationGraph",
        "type": "object"
      },
      "SemanticGraphNode": {
        "additionalProperties": true,
        "description": "A term on the semantic-layer exploration graph.",
        "properties": {
          "column_attributes_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column Attributes Count"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "relationship_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship Count"
          },
          "sql_attributes_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sql Attributes Count"
          },
          "synonyms": {
            "items": {
              "type": "string"
            },
            "title": "Synonyms",
            "type": "array"
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/ZoneChip"
            },
            "title": "Zones",
            "type": "array"
          }
        },
        "required": ["id"],
        "title": "SemanticGraphNode",
        "type": "object"
      },
...
      "TableExplorationDetails": {
        "additionalProperties": true,
        "description": "The table-detail modal: its SQL queries and one page of its Terms.",
        "properties": {
          "queries": {
            "items": {
              "$ref": "#/components/schemas/TableSqlQuery"
            },
            "title": "Queries",
            "type": "array"
          },
          "terms": {
            "items": {
              "$ref": "#/components/schemas/TermSummary"
            },
            "title": "Terms",
            "type": "array"
          },
          "terms_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Total"
          }
        },
        "title": "TableExplorationDetails",
        "type": "object"
      },
      "TableSqlQuery": {
        "additionalProperties": true,
        "description": "A stored SQL query that references the table.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "sql": {
            "title": "Sql",
            "type": "string"
          }
        },
        "required": ["id", "sql"],
        "title": "TableSqlQuery",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ZoneChip": {
        "additionalProperties": true,
        "description": "A zone as rendered next to something else it grants access to.\n\n``enabled`` is false for a zone whose ``Zone`` label was swapped for\n``disableZone``; those are returned to admins only.",
        "properties": {
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": ["id"],
        "title": "ZoneChip",
        "type": "object"
      },
...
    }
  }
}